aug19 commit
